pyproject.toml
[tool.poetry]
name = "ritual"
version = "0.1.0"
description = ""
authors = ["verghil <verghil@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
pytelegrambotapi = "^4.22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
